+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - GP2-3352 - remove feature FEATURE_FLAG_HARD_CODE_USER_INDUSTRIES_EXPERTISE
 - GP2-3352 - Obfuscate the export plan id in urls
 - GP2-3396 - Startdate Endate validation
+- GP2-3404 - fix ep dashboard default
 - GP2-3384 - Funding and credit: Amend pdf content
 
 ### Fixed bugs